Reorganize sample AreaDefinitions
